Treat <esi:...> tags as CDATA
The ESI specs (http://www.w3.org/TR/esi-lang for examle) describe the ESI language as an "in-markup XML-based language".

Yet, it does not say anything about encoding the special XML characters. Most notably, this is the "&" ampersand that may show up in esi:include URLs.

The problem is that ESI tags may be embedded in all different kinds of content, for example plain text files. In these cases, you probably won't process the file with a full-fledged XML parser and thus not having the ampersands encoded is not going to be a problem.

As we also need to assume that the ESI implementation in surrogate caches is rather based on simple pattern/text replacement than complete XML processing, there's no guarantee that such implementations will be able to treat "&amp;" sequences found in URLs as "&" characters.

So, the safest option seems to be to place <esi:...> tags in <![CDATA[...]]> sections prior to XML processing and strip that markup again when dumping the DOM back to XML.