Handle non-system DTDs gracefully
Unfortunately, there is no way to de-register the libxml entity loader.
So it may happen that our loader is called by "other" code that loads
and validates XML against a local schema definition.

In that case, the $system parameter remains unset and we need to
gracefully deal with it, avoiding a notice.